Rename: 폴더 구조 변경 (auth와 api 분리)
