+++ src/main/java/com/mamoori/mamooriback/api/controller/TestController.java
@@ -1,4 +1,4 @@
-package com.mamoori.mamooriback.controller;
+package com.mamoori.mamooriback.api.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
