+++ src/main/java/com/mamoori/mamooriback/api/entity/Comment.java
@@ -1,4 +1,4 @@
-package com.mamoori.mamooriback.entity;
+package com.mamoori.mamooriback.api.entity;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -7,8 +7,7 @@ import javax.persistence.*;
 
 @Entity
 @Getter @Setter
-@Table(name = "likes")
-public class Like extends BaseTimeEntity {
+public class Comment extends BaseTimeEntity {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
