+++ src/main/java/com/mamoori/mamooriback/api/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package com.mamoori.mamooriback.repository;
+package com.mamoori.mamooriback.api.repository;
 
-import com.mamoori.mamooriback.entity.Post;
+import com.mamoori.mamooriback.api.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
