+++ src/main/java/com/mamoori/mamooriback/api/repository/UserRepository.java
@@ -1,8 +1,8 @@
-package com.mamoori.mamooriback.repository;
+package com.mamoori.mamooriback.api.repository;
 
 import java.util.Optional;
 
-import com.mamoori.mamooriback.entity.User;
+import com.mamoori.mamooriback.api.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User,Long> {
