+++ src/main/java/com/mamoori/mamooriback/api/service/PostService.java
@@ -1,7 +1,7 @@
-package com.mamoori.mamooriback.service;
+package com.mamoori.mamooriback.api.service;
 
-import com.mamoori.mamooriback.controller.request.PostRequest;
-import com.mamoori.mamooriback.dto.PostResponse;
+import com.mamoori.mamooriback.api.controller.request.PostRequest;
+import com.mamoori.mamooriback.api.dto.PostResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
