+++ src/main/java/com/mamoori/mamooriback/api/service/UserService.java
@@ -1,8 +1,8 @@
-package com.mamoori.mamooriback.service;
+package com.mamoori.mamooriback.api.service;
 
 import java.util.Optional;
 
-import com.mamoori.mamooriback.entity.User;
+import com.mamoori.mamooriback.api.entity.User;
 
 public interface UserService {
 	Optional<User> findByEmail(String email);
