+++ src/main/java/com/mamoori/mamooriback/api/service/impl/UserServiceImpl.java
@@ -1,18 +1,17 @@
-package com.mamoori.mamooriback.service.impl;
+package com.mamoori.mamooriback.api.service.impl;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import com.mamoori.mamooriback.entity.User;
-import com.mamoori.mamooriback.repository.UserRepository;
+import com.mamoori.mamooriback.api.controller.request.UserRequest;
+import com.mamoori.mamooriback.api.service.UserService;
+import com.mamoori.mamooriback.api.entity.User;
+import com.mamoori.mamooriback.api.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import com.mamoori.mamooriback.controller.request.UserRequest;
-import com.mamoori.mamooriback.service.UserService;
 
 @Service
 public class UserServiceImpl implements UserService {
